+++ wordipy/main.py
@@ -5,16 +5,13 @@ Created on Thu Oct 23 22:11:38 2019
 
 @author: amit
 """
+from . import helpers
 import re
 try:
     from word2number import w2n
 except ImportError:
     print("Install word2number")
-try:
-    import nltk
-    from nltk.stem import WordNetLemmatizer #,PorterStemmer
-except ImportError:
-    print("Install nltk")
+
 try:
     import wordipy.vocabulary as voc
     pass
@@ -34,26 +31,6 @@ except ModuleNotFoundError:
 #import vocabulary as voc
 
 #global output
-
-def processing(io):
-    """
-    this function helps in preprocessing
-    """
-    input_str = io 
-    input_lower = input_str.lower()
-
-    words = input_lower.split(" ")
-    
-    #porter_stemmer=PorterStemmer()
-    #stemmed_words=[porter_stemmer.stem(word) for word in words]
-    
-    #nltk.download('wordnet')
-    
-    lemmatizer = WordNetLemmatizer()
-    lemmatized_words=[lemmatizer.lemmatize(word=word, pos='n') for word in words]
-    
-    process_words = lemmatized_words
-    return process_words
 
 
 def amountize(io):
@@ -76,17 +53,19 @@ def amountize(io):
     '$300'
     >>> amountize("five thousand three hundred and sixty one yen")
     '¥5361'
-    """    
-    process_words = processing(io)
+    """
+    process_words = helpers.processing(io)
     if len(set(process_words) & set(voc.currencies.keys())):
-        currency = set(process_words) & set(voc.currencies.keys()) #raise exception /try except
+        currency = set(process_words) & set(
+            voc.currencies.keys())  # raise exception /try except
         currency_symbol = voc.currencies[currency.pop()]
         amount = " ".join(process_words[:-1])
         amount_num = w2n.word_to_num(amount)
-        
+
         output_str = currency_symbol + str(amount_num)
         #print(output_str)
         return output_str
+
 
 def abbreviationize(input_str, sep=""):
     """
@@ -114,11 +93,11 @@ def abbreviationize(input_str, sep=""):
     >>> abbreviationize('Quadruple Baby', sep=",")
     'Baby, Baby, Baby, Baby' 
     """
-    process_words = processing(input_str)
+    process_words = helpers.processing(input_str)
     input_str_list = input_str.split(" ")
-    
+
     intersection = set(process_words) & set(voc.tuple_.keys())
-    
+
     #print(intersection)
     if len(intersection):
         '''
@@ -127,8 +106,9 @@ def abbreviationize(input_str, sep=""):
         times = intersection.pop()
         times_num = int(voc.tuple_[times])
         item = input_str.replace(times, '')[len(times):] + sep
-        output = (item*times_num).strip() # lstrip() can also be used, but strip trims spaces from both sides
-        last_char_remove = True if sep != ""  else False
+        # lstrip() can also be used, but strip trims spaces from both sides
+        output = (item*times_num).strip()
+        last_char_remove = True if sep != "" else False
         if last_char_remove:
             output = output[:-1]
         check = final_abbreviationize(output, input_str_list)
@@ -138,14 +118,15 @@ def abbreviationize(input_str, sep=""):
     else:
         return final_abbreviationize(input_str, input_str_list)
 
-def final_abbreviationize(output, input_str_list):            
+
+def final_abbreviationize(output, input_str_list):
     """ This is a helper function for abbreviationize"""
-    
-    if all([x.isupper() for x in input_str_list]) or all(x.isupper() for x in output.split()): 
+
+    if all([x.isupper() for x in input_str_list]) or all(x.isupper() for x in output.split()):
         '''
         if all are upper ex: C M -> CM
         '''
-        pattern = re.compile(r's+') 
+        pattern = re.compile(r's+')
         output = re.sub(pattern, '', output)
         #assert(len("".join(input_str_list)) == len(a))
         output = output.replace(" ", "")
